pref: 使前端支持i8n、暗亮主题跟随
