+++ pages/dashboard/index.html
@@ -1,9 +1,9 @@
 <!DOCTYPE html>
-<html lang="zh-CN">
+<html lang="en">
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no">
-    <title>亨利的战利品 - 表情包管理</title>
+    <title>Sticker Dashboard</title>
     <script src="https://cdn.jsdelivr.net/npm/vue@3.4.38/dist/vue.global.js"></script>
     <link rel="stylesheet" href="./app.css">
 </head>
